chore: restore model dropdown and add model info button
